Make it easier to deal with defaults
I think mapping t to tab is really confusing. I'm also not sure when using a new line as the separator in a TSV would ever make sense, so it's probably best to leave it as a tab. TBH, a TSV sort of is named by the usage of tabs, so maybe it's better to call it CSV if you want to use alternate separators