Update graph writer interface
- Abstract shared code between RDF and TSV writers into base class (OpenBioLinkGraphWriter)
- Abstract one level further (GraphWriter) to allow other types of writers that don't make the same kinds of assumptions (preparing for #29)

It was a bit tricky navigating around a lot of these classes - the combination of wrapping static variables in the class, then using static methods that instantiate the class that wraps static variables was hard to follow. I switched these to just using self, since the code that uses these writers makes an instance in the end.